+++ crates/symbolicator-service/src/caching/cache_key.rs
@@ -8,7 +8,6 @@ use crate::types::Scope;
 
 #[derive(Debug, Clone, Eq)]
 pub struct CacheKey {
-    legacy_cache_key: Arc<str>,
     metadata: Arc<str>,
     hash: [u8; 32],
 }
@@ -34,7 +33,9 @@ impl std::hash::Hash for CacheKey {
 impl CacheKey {
     /// Creates a [`CacheKey`] for the given [`RemoteFile`] tied to [`Scope`].
     pub fn from_scoped_file(scope: &Scope, file: &RemoteFile) -> Self {
-        Self::legacy_builder(scope, file).build()
+        let mut builder = Self::scoped_builder(scope);
+        builder.write_file_meta(file).unwrap();
+        builder.build()
     }
 
     /// Returns the human-readable metadata that forms the basis of the [`CacheKey`].
@@ -58,51 +59,18 @@ impl CacheKey {
         path
     }
 
-    /// Creates a stable relative path for the given [`RemoteFile`] tied to [`Scope`].
-    ///
-    /// This is being used as the "legacy" [`CacheKey`], and also forms the basis for the more
-    /// precise modern cache key.
-    pub fn relative_scoped_file(scope: &Scope, file: &RemoteFile) -> String {
-        let scope = safe_path_segment(scope.as_ref());
-        let cache_key = safe_path_segment(&file.cache_key());
-        format!("{scope}/{cache_key}")
-    }
-
     /// Create a [`CacheKeyBuilder`] that can be used to build a cache key consisting of all its
     /// contributing sources.
-    pub fn legacy_builder(scope: &Scope, file: &RemoteFile) -> CacheKeyBuilder {
-        let legacy_cache_key = Self::relative_scoped_file(scope, file);
+    pub fn scoped_builder(scope: &Scope) -> CacheKeyBuilder {
         let metadata = format!("scope: {scope}\n\n");
-
-        let mut builder = CacheKeyBuilder {
-            legacy_cache_key,
-            metadata,
-        };
-        builder.write_file_meta(file).unwrap();
-        builder
-    }
-
-    /// Returns the full cache path for this key inside the provided cache directory.
-    pub fn legacy_cache_path(&self, version: u32) -> String {
-        let mut path = if version != 0 {
-            format!("{version}/")
-        } else {
-            String::new()
-        };
-        path.push_str(&self.legacy_cache_key);
-        path
+        CacheKeyBuilder { metadata }
     }
 
     #[cfg(test)]
     pub fn for_testing(key: impl Into<String>) -> Self {
-        let legacy_cache_key = key.into();
-        let metadata = legacy_cache_key.clone();
+        let metadata = key.into();
 
-        CacheKeyBuilder {
-            legacy_cache_key,
-            metadata,
-        }
-        .build()
+        CacheKeyBuilder { metadata }.build()
     }
 }
 
@@ -113,7 +81,6 @@ impl CacheKey {
 /// This input in then being hashed to form the [`CacheKey`], and can also be serialized alongside
 /// the cache files to help debugging.
 pub struct CacheKeyBuilder {
-    legacy_cache_key: String,
     metadata: String,
 }
 
@@ -134,7 +101,6 @@ impl CacheKeyBuilder {
         let hash = <[u8; 32]>::try_from(hash).expect("sha256 outputs 32 bytes");
 
         CacheKey {
-            legacy_cache_key: self.legacy_cache_key.into(),
             metadata: self.metadata.into(),
             hash,
         }
@@ -145,14 +111,6 @@ impl fmt::Write for CacheKeyBuilder {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         self.metadata.write_str(s)
     }
-}
-
-/// Protect against:
-/// * ".."
-/// * absolute paths
-/// * ":" (not a threat on POSIX filesystems, but confuses OS X Finder)
-fn safe_path_segment(s: &str) -> String {
-    s.replace(['.', '/', '\\', ':'], "_")
 }
 
 #[cfg(test)]
@@ -179,8 +137,6 @@ mod tests {
 
         let key = CacheKey::from_scoped_file(&scope, &file);
 
-        assert_eq!(&key.legacy_cache_path(0), "global/foo_bar_baz");
-        assert_eq!(&key.legacy_cache_path(1), "1/global/foo_bar_baz");
         assert_eq!(
             &key.cache_path(0),
             "v0/6f/200788/bd4e6760d55bf6bd50c6d6e98b52379e194f9989fb788b4d37796427"
@@ -190,13 +146,12 @@ mod tests {
             "scope: global\n\nsource: foo\nlocation: file:///bar.baz\n"
         );
 
-        let builder = CacheKey::legacy_builder(&scope, &file);
-        let built_key = builder.build();
+        let built_key = CacheKey::from_scoped_file(&scope, &file);
 
-        assert_eq!(built_key.legacy_cache_path(0), key.legacy_cache_path(0));
         assert_eq!(built_key.cache_path(0), key.cache_path(0));
 
-        let mut builder = CacheKey::legacy_builder(&scope, &file);
+        let mut builder = CacheKey::scoped_builder(&scope);
+        builder.write_file_meta(&file).unwrap();
 
         let location = SourceLocation::new("bar.quux");
         let file = FilesystemRemoteFile::new(source, location).into();
@@ -204,7 +159,6 @@ mod tests {
         builder.write_file_meta(&file).unwrap();
         let key = builder.build();
 
-        assert_eq!(&key.legacy_cache_path(0), "global/foo_bar_baz");
         assert_eq!(
             &key.cache_path(0),
             "v0/07/e89036/d56878a462eb7949a744afa0a4deb5ed1b7a8154be16f7dd3b220518"