+++ crates/symbolicator-service/src/services/objects/meta_cache.rs
@@ -42,7 +42,7 @@ pub(super) struct FetchFileMetaRequest {
 /// Handle to local metadata file of an object.
 ///
 /// Having an instance of this type does not mean there is a downloaded object file behind
-/// it. We cache metadata separately (ObjectFeatures) because every symcache lookup requires
+/// it. We cache metadata separately ([`ObjectFeatures`]) because every SymCache lookup requires
 /// reading this metadata.
 #[derive(Clone, Debug)]
 pub struct ObjectMetaHandle {
@@ -58,7 +58,9 @@ impl ObjectMetaHandle {
     }
 
     pub fn cache_key_builder(&self) -> CacheKeyBuilder {
-        CacheKey::legacy_builder(&self.scope, &self.file_source)
+        let mut builder = CacheKey::scoped_builder(&self.scope);
+        builder.write_file_meta(&self.file_source).unwrap();
+        builder
     }
 
     pub fn features(&self) -> ObjectFeatures {
@@ -78,7 +80,7 @@ impl FetchFileMetaRequest {
     /// Fetches object file and derives metadata from it, storing this in the cache.
     ///
     /// This uses the data cache to fetch the requested file before parsing it and writing
-    /// the object metadata into the cache at `path`.  Technically the data cache could
+    /// the object metadata into `temp_file`.  Technically the data cache could
     /// contain the object file already but this is unlikely as normally the data cache
     /// expires before the metadata cache, so if the metadata needs to be re-computed then
     /// the data cache has probably also expired.
@@ -120,14 +122,8 @@ impl CacheItemRequest for FetchFileMetaRequest {
         Box::pin(self.compute_file_meta(temp_file))
     }
 
-    fn should_load(&self, data: &[u8]) -> bool {
-        serde_json::from_slice::<ObjectFeatures>(data).is_ok()
-    }
-
     /// Returns the [`ObjectMetaHandle`] at the given cache key.
     fn load(&self, data: ByteView<'static>) -> CacheEntry<Self::Item> {
-        // When CacheStatus::Negative we get called with an empty ByteView, for Malformed we
-        // get the malformed marker.
         let features = serde_json::from_slice(&data)?;
         Ok(Arc::new(ObjectMetaHandle {
             scope: self.scope.clone(),